+++ src/frontend/wizards/rename_encode.py
@@ -1,19 +1,26 @@
 from collections.abc import Sequence
 from pathlib import Path
 import re
-from typing import TYPE_CHECKING
+from typing import Any, TYPE_CHECKING
 
-from PySide6.QtCore import Qt, Slot
+from PySide6.QtCore import Qt, Signal, Slot
+from PySide6.QtGui import QCursor
 from PySide6.QtWidgets import (
+    QAbstractItemView,
     QCheckBox,
     QDialog,
+    QFrame,
     QGridLayout,
     QGroupBox,
     QHBoxLayout,
+    QHeaderView,
     QLabel,
     QLineEdit,
     QMessageBox,
+    QScrollArea,
     QSizePolicy,
+    QTableWidget,
+    QTableWidgetItem,
     QToolButton,
     QVBoxLayout,
     QWidget,
@@ -67,6 +74,7 @@ class RenameEncode(BaseWizardPage):
         self.backend = RenameEncodeBackEnd()
         self._input_ext: str | None = None
         self._token_window: QWidget | None = None
+        self._overridden_tokens = set()
 
         self.media_label = QLabel()
         self.media_label.setSizePolicy(
@@ -85,7 +93,9 @@ class RenameEncode(BaseWizardPage):
         options_group_box.setSizePolicy(
             QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Maximum
         )
-        options_layout = QGridLayout(options_group_box)
+
+        options_widget = QWidget(self)
+        options_layout = QGridLayout(options_widget)
         options_layout.setColumnStretch(0, 1)
         options_layout.setColumnStretch(1, 1)
         options_layout.setColumnStretch(2, 1)
@@ -93,7 +103,10 @@ class RenameEncode(BaseWizardPage):
 
         edition_lbl = QLabel("Edition", self)
         self.edition_combo = CustomComboBox(
-            completer=True, completer_strict=False, parent=self
+            completer=True,
+            completer_strict=False,
+            disable_mouse_wheel=True,
+            parent=self,
         )
         self._update_combo_box(self.edition_combo, EDITION_INFO)
         self.edition_combo.currentIndexChanged.connect(self._update_edition_combo)
@@ -103,25 +116,28 @@ class RenameEncode(BaseWizardPage):
         edition_combo_line_edit.editingFinished.connect(self._update_edition_combo)
 
         frame_size_lbl = QLabel("Frame Size", self)
-        self.frame_size_combo = CustomComboBox(parent=self)
+        self.frame_size_combo = CustomComboBox(disable_mouse_wheel=True, parent=self)
         self._update_combo_box(self.frame_size_combo, FRAME_SIZE_INFO)
         self.frame_size_combo.currentIndexChanged.connect(self._update_frame_size_combo)
 
         localization_lbl = QLabel("Localization", self)
-        self.localization_combo = CustomComboBox(parent=self)
+        self.localization_combo = CustomComboBox(disable_mouse_wheel=True, parent=self)
         self._update_combo_box(self.localization_combo, LOCALIZATION_INFO)
         self.localization_combo.currentIndexChanged.connect(
             self._update_localization_combo
         )
 
         re_release_lbl = QLabel("Rerelease", self)
-        self.re_release_combo = CustomComboBox(parent=self)
+        self.re_release_combo = CustomComboBox(disable_mouse_wheel=True, parent=self)
         self._update_combo_box(self.re_release_combo, RE_RELEASE_INFO)
         self.re_release_combo.currentIndexChanged.connect(self._update_re_release_combo)
 
         self.repack_reason_lbl = QLabel("Repack Reason", self)
         self.repack_reason_combo = CustomComboBox(
-            completer=True, completer_strict=False, parent=self
+            completer=True,
+            completer_strict=False,
+            disable_mouse_wheel=True,
+            parent=self,
         )
         self.repack_reason_combo.setSizeAdjustPolicy(
             CustomComboBox.SizeAdjustPolicy.AdjustToMinimumContentsLengthWithIcon
@@ -136,7 +152,10 @@ class RenameEncode(BaseWizardPage):
 
         self.proper_reason_lbl = QLabel("Proper Reason", self)
         self.proper_reason_combo = CustomComboBox(
-            completer=True, completer_strict=False, parent=self
+            completer=True,
+            completer_strict=False,
+            disable_mouse_wheel=True,
+            parent=self,
         )
         self.proper_reason_combo.setSizeAdjustPolicy(
             CustomComboBox.SizeAdjustPolicy.AdjustToMinimumContentsLengthWithIcon
@@ -158,12 +177,25 @@ class RenameEncode(BaseWizardPage):
             (self.re_release_combo, RE_RELEASE_INFO),
         )
 
+        self.remux_checkbox = QCheckBox("REMUX", self)
+        self.remux_checkbox.setToolTip("Toggle REMUX token")
+        self.remux_checkbox.toggled.connect(self._remux_toggle)
+
+        self.hybrid_checkbox = QCheckBox("HYBRID", self)
+        self.hybrid_checkbox.setToolTip("Toggle HYBRID token")
+        self.hybrid_checkbox.toggled.connect(self._hybrid_toggle)
+
+        checkboxes_layout = QHBoxLayout()
+        checkboxes_layout.setContentsMargins(0, 0, 0, 0)
+        checkboxes_layout.addWidget(self.remux_checkbox)
+        checkboxes_layout.addWidget(self.hybrid_checkbox)
+
         release_group_lbl = QLabel("Release Group", self)
         self.release_group_entry = QLineEdit(self)
         self.release_group_entry.setToolTip("Release group name")
         self.release_group_entry.textEdited.connect(self.update_generated_name)
 
-        token_override_lbl = QLabel("Override Movie Name Token", self)
+        token_override_lbl = QLabel("Override File Name Tokens", self)
         view_tokens_popup_btn: QToolButton = build_auto_theme_icon_buttons(
             QToolButton,
             "token.svg",
@@ -183,26 +215,18 @@ class RenameEncode(BaseWizardPage):
         self.token_override = QLineEdit(self)
         self.token_override.textEdited.connect(self.update_generated_name)
 
+        self.rename_token_control = RenameTokenControl(self)
+        self.rename_token_control.row_modified.connect(self._update_override)
+
         self.override_group = QGroupBox(
             title="Override", parent=self, checkable=True, checked=False
         )
+        self.override_group.toggled.connect(self._on_override_group_toggled)
         self.override_group.toggled.connect(self.update_generated_name)
         override_group_layout = QVBoxLayout(self.override_group)
         override_group_layout.addLayout(token_override_layout)
         override_group_layout.addWidget(self.token_override)
-
-        self.remux_checkbox = QCheckBox("REMUX", self)
-        self.remux_checkbox.setToolTip("Toggle REMUX token")
-        self.remux_checkbox.toggled.connect(self._remux_toggle)
-
-        self.hybrid_checkbox = QCheckBox("HYBRID", self)
-        self.hybrid_checkbox.setToolTip("Toggle HYBRID token")
-        self.hybrid_checkbox.toggled.connect(self._hybrid_toggle)
-
-        checkboxes_layout = QHBoxLayout()
-        checkboxes_layout.setContentsMargins(0, 0, 0, 0)
-        checkboxes_layout.addWidget(self.remux_checkbox)
-        checkboxes_layout.addWidget(self.hybrid_checkbox)
+        override_group_layout.addWidget(self.rename_token_control)
 
         options_layout.addWidget(edition_lbl, 0, 0)
         options_layout.addWidget(self.edition_combo, 1, 0)
@@ -223,6 +247,14 @@ class RenameEncode(BaseWizardPage):
         options_layout.addWidget(build_h_line((6, 4, 6, 4)), 21, 0, 1, 3)
         options_layout.addWidget(self.override_group, 22, 0, 1, 3)
 
+        self.options_scroll_area = QScrollArea(self, widgetResizable=True)
+        self.options_scroll_area.setFrameShape(QFrame.Shape.NoFrame)
+        self.options_scroll_area.setWidget(options_widget)
+
+        group_box_layout = QVBoxLayout(options_group_box)
+        group_box_layout.setContentsMargins(0, 0, 0, 0)
+        group_box_layout.addWidget(self.options_scroll_area)
+
         output_group_box = QGroupBox("Output")
         output_group_box.setSizePolicy(
             QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Maximum
@@ -230,6 +262,7 @@ class RenameEncode(BaseWizardPage):
         output_layout = QVBoxLayout(output_group_box)
         self.output_entry = QLineEdit()
         self.output_entry.setToolTip("Suggested name, updates automatically")
+        self.output_entry.setReadOnly(True)
 
         output_layout.addWidget(self.output_entry)
 
@@ -303,6 +336,24 @@ class RenameEncode(BaseWizardPage):
 
             return True
         return False
+
+    @Slot(bool)
+    def _on_override_group_toggled(self, checked: bool) -> None:
+        if not checked:
+            # remove only the overridden tokens
+            for k in self._overridden_tokens:
+                self.backend.override_tokens.pop(k, None)
+                self.config.shared_data.dynamic_data.get("override_tokens", {}).pop(
+                    k, None
+                )
+            self._overridden_tokens.clear()
+            self.update_generated_name()
+
+    @Slot(tuple)
+    def _update_override(self, data: tuple[str, str]) -> None:
+        self.backend.override_tokens[data[0]] = data[1]
+        self._overridden_tokens.add(data[0])
+        self.update_generated_name()
 
     @Slot()
     def _see_tokens(self) -> None:
@@ -490,10 +541,20 @@ class RenameEncode(BaseWizardPage):
             movie_clean_title_rules=self.config.cfg_payload.mvr_clean_title_rules,
         )
 
-        # update the output entry
-        if get_file_name:
+        if get_file_name and self.backend.token_replacer:
+            # update rename token control
+            sort_token_order = re.findall(r"\{(.*?)\}", token)
+            sort_token_data = self.backend.token_replacer.token_data.get_dict()  # pyright: ignore[reportAttributeAccessIssue]
+            sorted_token_data = {
+                k: sort_token_data[k]
+                for k in sort_token_order
+                if k in sort_token_data and sort_token_data[k]
+            }
+            self.rename_token_control.populate_table(sorted_token_data)
+
+            # update entries
             self._input_ext = get_file_name.suffix
-            self.output_entry.setText(str(get_file_name))
+            self.output_entry.setText(str(get_file_name.with_suffix("")))
 
     def _reset_re_release_reason_widgets(self) -> None:
         """Hide and reset both repack and proper reason widgets."""
@@ -501,14 +562,12 @@ class RenameEncode(BaseWizardPage):
             (self.repack_reason_lbl, self.repack_reason_combo),
             (self.proper_reason_lbl, self.proper_reason_combo),
         ]:
-            combo.blockSignals(True)
             lbl.hide()
             combo.hide()
             combo.setCurrentIndex(0)
             line_edit = combo.lineEdit()
             if line_edit:
                 line_edit.setPlaceholderText(self.REASON_STR)
-            combo.blockSignals(False)
 
     def _enable_re_release_widgets(self, combo: CustomComboBox) -> None:
         """Show the appropriate reason widgets based on rerelease combo selection."""
@@ -531,13 +590,17 @@ class RenameEncode(BaseWizardPage):
             self.localization_combo,
             self.re_release_combo,
         ):
+            combo_box.blockSignals(True)
             combo_box.setCurrentIndex(0)
+            combo_box.blockSignals(False)
         self._reset_re_release_reason_widgets()
         self.release_group_entry.clear()
-        self._input_ext = None
+        self.options_scroll_area.verticalScrollBar().setValue(0)
         self.output_entry.clear()
 
+        self._input_ext = None
         self._close_token_window()
+        self._overridden_tokens.clear()
 
         self.backend.reset()
 
@@ -549,3 +612,108 @@ class RenameEncode(BaseWizardPage):
         combobox.addItem("")
         for item in items:
             combobox.addItem(item[0])
+
+
+class RenameTokenControl(QWidget):
+    row_modified = Signal(tuple)
+
+    def __init__(self, parent=None) -> None:
+        super().__init__(parent)
+
+        desc = QLabel(
+            """\
+            <br>
+            <span style="font-style: italic; font-size: smaller;"><strong>Note:</strong> Modifying the tokens 
+            below will only update corresponding <strong>title</strong> tokens (global/per tracker) if they share 
+            the same token.</span>""",
+            self,
+            wordWrap=True,
+        )
+
+        self.search_bar = QLineEdit(self)
+        self.search_bar.setPlaceholderText("Search...")
+        self.search_bar.textChanged.connect(self.filter_table)
+
+        self.table = QTableWidget(self)
+        self.table.setColumnCount(2)
+        self.table.setMinimumHeight(180)
+        self.table.setHorizontalHeaderLabels(("Token", "Value (click to edit)"))
+        self.table.setFrameShape(QFrame.Shape.Box)
+        self.table.setFrameShadow(QFrame.Shadow.Sunken)
+        self.table.setCursor(QCursor(Qt.CursorShape.PointingHandCursor))
+        self.table.itemChanged.connect(self._item_changed)
+
+        self.main_layout = QVBoxLayout(self)
+        self.main_layout.setContentsMargins(0, 0, 0, 0)
+        self.main_layout.addWidget(desc)
+        self.main_layout.addWidget(self.search_bar)
+        self.main_layout.addWidget(self.table)
+
+    def populate_table(self, tokens: dict[str, Any]) -> None:
+        self.table.blockSignals(True)
+        self.table.clearContents()
+        self.table.setRowCount(len(tokens))
+
+        for idx, (token, value) in enumerate(tokens.items()):
+            # build token item
+            token_item = QTableWidgetItem(f"{{{token}}}")
+            token_item.setFlags(
+                token_item.flags()
+                & ~Qt.ItemFlag.ItemIsEditable
+                & ~Qt.ItemFlag.ItemIsSelectable
+            )
+            self.table.setItem(idx, 0, token_item)
+
+            # build editable item
+            item = QTableWidgetItem(value)
+            item.setFlags(
+                item.flags()
+                | Qt.ItemFlag.ItemIsEditable
+                | Qt.ItemFlag.ItemIsSelectable
+                | Qt.ItemFlag.ItemIsEnabled
+            )
+            self.table.setItem(idx, 1, item)
+
+        self.setup_table_properties()
+        self.table.blockSignals(False)
+
+    def setup_table_properties(self) -> None:
+        self.table.resizeColumnsToContents()
+        self.table.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeMode.Fixed)
+        self.table.horizontalHeader().setStretchLastSection(True)
+        self.table.verticalHeader().setVisible(False)
+        self.table.setSelectionMode(QTableWidget.SelectionMode.NoSelection)
+        self.table.setHorizontalScrollMode(QAbstractItemView.ScrollMode.ScrollPerPixel)
+        self.table.setAutoScroll(False)
+
+    @Slot(QTableWidgetItem)
+    def _item_changed(self, item: QTableWidgetItem) -> None:
+        token = self.table.item(item.row(), 0)
+        if token and item:
+            self.row_modified.emit((token.text().strip("{}"), item.text()))
+
+    def get_token_values(self) -> dict:
+        """Return a dict of token: value"""
+        values = {}
+        for row in range(self.table.rowCount()):
+            token = self.table.item(row, 0)
+            value = self.table.item(row, 1)
+            if token is not None and value is not None:
+                values[token.text()] = value.text()
+        return values
+
+    @Slot(str)
+    def filter_table(self, text: str) -> None:
+        for row in range(self.table.rowCount()):
+            match = False
+            for column in range(self.table.columnCount()):
+                item = self.table.item(row, column)
+                if item and text.lower() in item.text().lower():
+                    match = True
+                    break
+            self.table.setRowHidden(row, not match)
+
+    def reset(self) -> None:
+        self.table.blockSignals(True)
+        self.table.clearContents()
+        self.table.setAutoScroll(False)